+++ CHANGELOG.markdown
@@ -3,6 +3,7 @@
 ## latest
 
 * add --changelog option
+* remove completion command, cache aliases in a file instead, closes #51
 
 ## v0.4.11 2014-01-25